Add WebsocketServer.AllowOrigins() to specify allowed origins
The AllowOrigins() function configures the websocket server with origin checking logic that allows origins based on glob matching.  This handles most origin matching logic that a router needs, so is provided as a utility so that router implementers do not have to supply their own, likely similar, logic.

If all origins are allowed, specified by a wildcard glob "*", then no actual matching is performed and the origin check simply returns true.  Exact matches are checked first to avoid the possibly unneeded overhead of glob matching.

The websocket server still allows custom origin check logic to be assigned, by assigning a check function to WebsocketServer.Upgrader.CheckOrigin.

This fixes #126
