+++ router/websocketserver.go
@@ -9,6 +9,9 @@ import (
 	"io"
 	"net"
 	"net/http"
+	"net/url"
+	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/gammazero/nexus/stdlog"
@@ -35,19 +38,10 @@ type protocol struct {
 //
 // Web browsers allow Javascript applications to open a WebSocket connection to
 // any host.  It is up to the server to enforce an origin policy using the
-// Origin request header sent by the browser.
-//
-// The WebsocketServer calls the function specified in the
-// WebsocketServer.Upgrader.CheckOrigin field to check the origin.  If the
-// CheckOrigin function returns false, then the WebSocket handshake fails with
-// HTTP status 403.  To supply a CheckOrigin function:
-//
-//     s := NewWebsocketServer(r)
-//     s.Upgrader.CheckOrigin = func(r *http.Request) bool { ... }
-//
-// If the CheckOrigin field is nil, then a safe default is used: fail
-// the handshake if the Origin request header is present and the Origin host is
-// not equal to the Host request header.
+// Origin request header sent by the browser.  To specify origins allowed by
+// the server, call AllowOrigins() to allow origins matching glob patterns, or
+// assign a custom function to the server's Upgrader.CheckOrigin.  See
+// AllowOrigins() for details.
 type WebsocketServer struct {
 	// Upgrader specifies parameters for upgrading an HTTP connection to a
 	// websocket connection.  See:
@@ -140,6 +134,84 @@ func NewWebsocketServer(r Router) *WebsocketServer {
 		&serialize.CBORSerializer{})
 
 	return s
+}
+
+// AllowOrigins configures the server to allow origins that match the specified
+// glob patterns.  If the origin in the websocket upgrade request is not set,
+// is equal to the request host, or matches one of the allowed patterns, then
+// the websocket connection is allowed.  A pattern is in the form of a shell
+// glob as described here: https://golang.org/pkg/path/filepath/#Match
+//
+// For example:
+//   s := NewWebsocketServer(r)
+//   s.AllowOrigins([]string{"*.domain.com", "*.domain.net"})
+//
+// To allow all origins, specify a wildcard only:
+//   s.AllowOrigins([]string{"*"})
+//
+// Origins with Ports
+//
+// To allow origins that have a port number, specify the port as part of the
+// origin pattern, or use a wildcard to match the ports.
+//
+// Allow a specific port, by specifying the expected port number:
+//   s.AllowOrigins([]string{"*.somewhere.com:8080"})
+//
+// Allow individual ports, by specifying each port:
+//  err := s.AllowOrigins([]string{
+//      "*.somewhere.com:8080",
+//      "*.somewhere.com:8905",
+//      "*.somewhere.com:8908",
+//  })
+//
+// Allow any port, by specifying a wildcard. Be sure to include the colon ":"
+// so that the pattern does not match a longer origin.  Without the ":" this
+// example would also match "x.somewhere.comics.net"
+//   err := s.AllowOrigins([]string{"*.somewhere.com:*"})
+//
+// Custom Origin Checks
+//
+// Alternatively, a custom function my be configured to supplied any origin
+// checking logic your application needs.  The WebsocketServer calls the
+// function specified in the WebsocketServer.Upgrader.CheckOrigin field to
+// check the origin.  If the CheckOrigin function returns false, then the
+// WebSocket handshake fails with HTTP status 403.  To supply a CheckOrigin
+// function:
+//
+//     s := NewWebsocketServer(r)
+//     s.Upgrader.CheckOrigin = func(r *http.Request) bool { ... }
+//
+// Default Behavior
+//
+// If AllowOrigins() is not called, and Upgrader.CheckOrigin is nil, then a
+// safe default is used: fail the handshake if the Origin request header is
+// present and the Origin host is not equal to the Host request header.
+func (s *WebsocketServer) AllowOrigins(origins []string) error {
+	if len(origins) == 0 {
+		return nil
+	}
+	var exacts, globs []string
+	for _, o := range origins {
+		// If allowing any origins, then return simple "true" function.
+		if o == "*" {
+			s.Upgrader.CheckOrigin = func(r *http.Request) bool { return true }
+			return nil
+		}
+
+		// Do exact matching whenever possible, since it is more efficient.
+		if strings.ContainsAny(o, "*?[]^") {
+			if _, err := filepath.Match(o, o); err != nil {
+				return fmt.Errorf("error allowing origin, %s: %s", err, o)
+			}
+			globs = append(globs, strings.ToLower(o))
+		} else {
+			exacts = append(exacts, o)
+		}
+	}
+	s.Upgrader.CheckOrigin = func(r *http.Request) bool {
+		return checkOrigin(exacts, globs, r)
+	}
+	return nil
 }
 
 // Deprecated: Set WebsocketServer.Upgrader and WebsockServer.Xxx members
@@ -243,8 +315,8 @@ func (s *WebsocketServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	// If request capture is enabled, then the HTTP upgrade http.Request in the
-	// HELLO and session details as transport.auth details.request.
+	// If request capture is enabled, then save the HTTP upgrade http.Request
+	// in the HELLO and session details as transport.auth details.request.
 	if s.EnableRequestCapture {
 		if authDict == nil {
 			authDict = wamp.Dict{}
@@ -307,4 +379,36 @@ func (s *WebsocketServer) handleWebsocket(conn *websocket.Conn, transportDetails
 	if err := s.router.AttachClient(peer, transportDetails); err != nil {
 		s.log.Println("Error attaching to router:", err)
 	}
+}
+
+// checkOrigin returns true if the origin is not set, is equal to the request
+// host, or matches one of the allowed patterns.  This function is assigned to
+// the server's Upgrader.CheckOrigin when AllowOrigins() is called.
+func checkOrigin(exacts, globs []string, r *http.Request) bool {
+	origin := r.Header["Origin"]
+	if len(origin) == 0 {
+		return true
+	}
+	u, err := url.Parse(origin[0])
+	if err != nil {
+		return false
+	}
+	if strings.EqualFold(u.Host, r.Host) {
+		return true
+	}
+
+	for i := range exacts {
+		if strings.EqualFold(u.Host, exacts[i]) {
+			return true
+		}
+	}
+	if len(globs) != 0 {
+		host := strings.ToLower(u.Host)
+		for i := range globs {
+			if ok, _ := filepath.Match(globs[i], host); ok {
+				return true
+			}
+		}
+	}
+	return false
 }
